担当分のnorm対応をしました。

+++ includes/vec3.h
@@ -6,7 +6,7 @@
 /*   By: htsutsum <htsutsum@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/11 18:29:29 by hayato            #+#    #+#             */
-/*   Updated: 2026/02/14 11:14:03 by htsutsum         ###   ########.fr       */
+/*   Updated: 2026/02/17 03:43:59 by htsutsum         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,15 +17,15 @@
 
 typedef struct s_vec3
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_vec3;
 
-t_vec3	vec3_init(float x, float y, float z);
+t_vec3	vec3_init(double x, double y, double z);
 t_vec3	vec3_add(t_vec3 a, t_vec3 b);
 t_vec3	vec3_sub(t_vec3 a, t_vec3 b);
-t_vec3	vec3_scale(t_vec3 v, float scalar);
+t_vec3	vec3_scale(t_vec3 v, double scalar);
 double	vec3_dot(t_vec3 a, t_vec3 b);
 t_vec3	vec3_cross(t_vec3 a, t_vec3 b);
 double	vec3_norm(t_vec3 v);
