+++ srcs/vec3_1.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   vec3_1.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hayato <hayato@student.42.fr>              +#+  +:+       +#+        */
+/*   By: htsutsum <htsutsum@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/11 17:57:14 by hayato            #+#    #+#             */
-/*   Updated: 2026/01/11 18:18:54 by hayato           ###   ########.fr       */
+/*   Updated: 2026/02/17 03:44:23 by htsutsum         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vec3.h"
 
-t_vec3	vec3_init(float x, float y, float z)
+t_vec3	vec3_init(double x, double y, double z)
 {
 	t_vec3	vec;
 
@@ -32,7 +32,7 @@ t_vec3	vec3_sub(t_vec3 a, t_vec3 b)
 	return (vec3_init(a.x - b.x, a.y - b.y, a.z - b.z));
 }
 
-t_vec3	vec3_scale(t_vec3 v, float scalar)
+t_vec3	vec3_scale(t_vec3 v, double scalar)
 {
 	return (vec3_init(v.x * scalar, v.y * scalar, v.z * scalar));
 }
